Move email template management to main frontend
