+++ services/headless-lms/src/controllers/cms/course_instances.rs
@@ -1,13 +1,6 @@
 //! Controllers for requests starting with `/api/v0/cms/course-instances`.
 
-use crate::{
-    controllers::ControllerResult,
-    domain::authorization::AuthUser,
-    models::{
-        course_instances::CourseInstance,
-        email_templates::{EmailTemplate, EmailTemplateNew},
-    },
-};
+use crate::{controllers::ControllerResult, models::course_instances::CourseInstance};
 use actix_web::web::ServiceConfig;
 use actix_web::web::{self, Json};
 use sqlx::PgPool;
@@ -49,39 +42,6 @@ async fn get_course_instance(
     Ok(Json(course_intance))
 }
 
-#[instrument(skip(payload, pool))]
-async fn post_new_email_template(
-    request_course_instance_id: web::Path<Uuid>,
-    payload: web::Json<EmailTemplateNew>,
-    pool: web::Data<PgPool>,
-    user: AuthUser,
-) -> ControllerResult<Json<EmailTemplate>> {
-    let mut conn = pool.acquire().await?;
-    let new_email_template = payload.0;
-    let email_template = crate::models::email_templates::insert_email_template(
-        &mut conn,
-        *request_course_instance_id,
-        new_email_template,
-        None,
-    )
-    .await?;
-    Ok(Json(email_template))
-}
-
-#[instrument(skip(pool))]
-async fn get_email_templates_by_course_instance_id(
-    request_course_instance_id: web::Path<Uuid>,
-    pool: web::Data<PgPool>,
-    user: AuthUser,
-) -> ControllerResult<Json<Vec<EmailTemplate>>> {
-    let mut conn = pool.acquire().await?;
-
-    let email_templates =
-        crate::models::email_templates::get_email_templates(&mut conn, *request_course_instance_id)
-            .await?;
-    Ok(Json(email_templates))
-}
-
 /**
 Add a route for each controller in this module.
 
@@ -90,13 +50,5 @@ The name starts with an underline in order to appear before other functions in t
 We add the routes by calling the route method instead of using the route annotations because this method preserves the function signatures for documentation.
 */
 pub fn _add_course_instances_routes(cfg: &mut ServiceConfig) {
-    cfg.route("/{course_instance_id}", web::get().to(get_course_instance))
-        .route(
-            "/{course_instance_id}/email-templates",
-            web::post().to(post_new_email_template),
-        )
-        .route(
-            "/{course_instance_id}/email-templates",
-            web::get().to(get_email_templates_by_course_instance_id),
-        );
+    cfg.route("/{course_instance_id}", web::get().to(get_course_instance));
 }
