+++ services/main-frontend/src/components/lists/CourseInstancesList.tsx
@@ -1,3 +1,4 @@
+import Link from "next/link"
 import React from "react"
 import { useQuery } from "react-query"
 
@@ -26,7 +27,14 @@ const CourseInstancesList: React.FC<CourseInstancesListProps> = ({ courseId }) =
         return (
           <li key={instance.id}>
             {instance.name ?? "Default"}{" "}
-            <a href={`/cms/course-instances/${instance.id}/manage-emails`}>Manage emails</a>
+            <Link
+              href={{
+                pathname: "/manage/course-instances/[id]/emails",
+                query: { id: instance.id },
+              }}
+            >
+              Manage emails
+            </Link>
           </li>
         )
       })}
