+++ services/main-frontend/src/services/backend/course-instances/index.ts
@@ -1,0 +1,28 @@
+import { EmailTemplate, EmailTemplateNew } from "../../../shared-module/bindings"
+import { mainFrontendClient } from "../../mainFrontendClient"
+
+export const postNewEmailTemplateForCourseInstance = async (
+  courseInstanceId: string,
+  data: EmailTemplateNew,
+): Promise<EmailTemplate> => {
+  const response = await mainFrontendClient.post(
+    `/course-instances/${courseInstanceId}/email-templates`,
+    data,
+    {
+      headers: { "Content-Type": "application/json" },
+    },
+  )
+  return response.data
+}
+
+export const fetchCourseInstanceEmailTemplates = async (
+  courseInstanceId: string,
+): Promise<EmailTemplate[]> => {
+  const response = await mainFrontendClient.get(
+    `/course-instances/${courseInstanceId}/email-templates`,
+    {
+      responseType: "json",
+    },
+  )
+  return response.data
+}
