+++ services/main-frontend/src/services/backend/email-templates/index.ts
@@ -1,0 +1,7 @@
+import { EmailTemplate } from "../../../shared-module/bindings"
+import { mainFrontendClient } from "../../mainFrontendClient"
+
+export const deleteEmailTemplate = async (id: string): Promise<EmailTemplate> => {
+  const response = await mainFrontendClient.delete(`/email-templates/${id}`)
+  return response.data
+}
